Initial cleanup (#5)
* Renamed vars for readability
* small improvements and cleanup
* Moving of BUILD files
* Moving string_links option
* Formatting

Preparation commit to move all py_libraries to their own BUILD files.
This solution at the moment is WIP as the pre-fix needed is unacceptable.
But this will serve as a base to be able to test if further changes work

Moved BUILD files to separate folders for better Separation
Cleaned up some smaller stuff
Conf.py is cleaner again