✨ Introduce needs-config-writer
https://needs-config-writer.useblocks.com

This greatly helps with distributed configs.
Also dynamic and static configuration can be merged.

And user of docs will get a ubproject.toml generated into
their confdir. This is a declarative duplicate of all the dynamically generated Sphinx-Needs configuration.
It makes it possible to use tools such as ubCode or ubc without any further setup.